chore: seperate multiplier and keccak proof
